Version bump to 2.1.0-volders
We need to enable the source map generation, but doing so we had this
issue (https://github.com/ember-cli/ember-cli-uglify/issues/4).

To solve it, we need to update the ember-intl to at least version 2.1.0.
The problem is that when updating it we start getting a new error:

```
node_modules/ember-cli-uglify/index.js:6
  included(app) {
          ^
Unexpected token (
```

This is because we are using an old version of node that not deal with
new javascript syntax very well.

So, to fix the issue without the need to the node version right
now, we created this version with this commit on top of the
version 2.1.0.